+++ aoeu/src/main/java/ua/laposhko/hmt/service/usercity/IUserCityService.java
@@ -9,5 +9,5 @@ import java.util.List;
  * Created by Sergey on 27.05.2015.
  */
 public interface IUserCityService extends GenericManager<UserCity> {
-    List<UserCity> findByCityId(int cityId);
+    List<UserCity> findByCityId(long cityId);
 }
